requirements/prod.txt
Django>=1.11
psycopg2==2.7
django-admin-tools
feedparser
raven
pip
django-bootstrap3
python-gitlab
django-simple-history
celery
jira
certifi
urllib3[secure]
django-extensions
redis
django-celery-beat
django-celery-results
gunicorn
setproctitle
six
flower
python-jenkins==0.4.14
requests==2.17.3
djangorestframework==3.6.3
